requirements.txt
Datetime==4.3
flask==1.1.2
insightface==0.1.5
keras==2.4.3
mtcnn==0.1.0
mxnet==1.2.0
numpy==1.18.5
opencv-python==4.4.0.46
psycopg2==2.8.6
pybase64==1.0.1
Pillow==7.2.0
requests==2.24.0
scikit-image==0.17.2
tensorflow==2.2.1
tqdm==4.54.1
